requirements.txt
Flask
nltk
transformers
torch
PyPDF2
python-docx
rouge-score
beautifulsoup4
requests
sentencepiece
fpdf
